1st running version, 1 LED, no interrupt
This is the first running version of the rgb flashlight program.
Reads button on PINB0 and outputs PWM to all pins on Port C.

This version currently doesn't support individually controlled LEDs
on Port C.

The decrement of the counter has to be controlled by Timer_0 interrupt
in following versions.